backend/requirements.txt
# FastAPI Backend Dependencies for RouteWhisper

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database - PostgreSQL (Structured Data)
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Database - MongoDB (Flexible Content)
pymongo==4.6.0
motor==3.3.2  # Async MongoDB driver
beanie==1.24.0  # ODM for MongoDB

# Database - Redis (Caching)
redis==5.0.1
aioredis==2.0.1  # Async Redis client

# HTTP Client & APIs
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Geographic Data
geopy==2.4.0
shapely==2.0.2

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1 